Envoi des exercices 1 à 4 car 5 en cours

+++ src/main/java/com/ipiecoles/java/java220/Entreprise.java
@@ -19,6 +19,6 @@ public class Entreprise {
      */
 	public static Double primeAnnuelleBase() {
     	Integer anneeEnCours = new LocalDate().getYear();
-    	return anneeEnCours / 2d;
+    	return (Double) (anneeEnCours / 2d);
     }
 }
